+++ CHANGES.rst
@@ -1,5 +1,6 @@
-v0.7.18, 2-9-2017
-    * Added performance improvements
+v0.7.25, 03-03-2017
+    * replaced xmltodict with an lxml-based XML to dict parser for higher performance
+    * get_all_objects now optionally returns the last-changed timestamp from listvs
 v0.7.0, 10-8-2016
     * Added Access Control, Template, Rule, SSO, FQDN, Site, Cluster, Range, CustomLocation, Certificate and Interface objects. Added more Virtual Service parameters.
 v0.6.19, 23-6-2016